+++ docs/changelog.md
@@ -3,3 +3,9 @@
 - Ajout système Git (repo initialisé + push GitHub)
 - Mise en place GridOverlay (shadow circles)
 - Fix synchronisation repo GitHub
+
+## 23/05/2026
+
+- Mise en place GridOverlay (shadow circles)
+- Passage mode exploration à mode combat
+- Grille